# sstable: vectorize shared prefix byte comparison